fix(parameters): validate reconfigure assignments in parameters layer
